Refactor chatQueue to use roomid
The chatQueue previously put a reference to the room directly in the
chatQueue array. This refactors it to use a room ID.

This mostly doesn't matter much, but it's nice not to have references
to rooms outside of the room, and makes for slightly cleaner code by
not needing the weird workaround to see if the room has expired.